+++ imio/dms/mail/subscribers.py
@@ -22,6 +22,7 @@ from imio.dms.mail import IM_READER_SERVICE_FUNCTIONS
 from imio.dms.mail.browser.settings import IImioDmsMailConfig
 from imio.dms.mail.interfaces import IActionsPanelFolder
 from imio.dms.mail.interfaces import IActionsPanelFolderAll
+from imio.dms.mail.interfaces import IActionsPanelFolderOnlyAdd
 from imio.dms.mail.interfaces import IPersonnelContact
 from imio.dms.mail.utils import IdmUtilsMethods
 from imio.dms.mail.utils import separate_fullname
@@ -350,7 +351,7 @@ def contact_plonegroup_change(event):
             full_title = obj.get_full_title(separator=' - ', first_index=1)
             if uid not in om_folder:
                 folder = api.content.create(container=om_folder, type='Folder', id=uid, title=full_title)
-                alsoProvides(folder, IActionsPanelFolder)
+                alsoProvides(folder, IActionsPanelFolderOnlyAdd)
                 alsoProvides(folder, INextPrevNotNavigable)
                 roles = ['Reader']
                 if registry['imio.dms.mail.browser.settings.IImioDmsMailConfig.org_templates_encoder_can_edit']:
@@ -362,7 +363,7 @@ def contact_plonegroup_change(event):
                     api.content.copy(source=base_model, target=folder)
             if uid not in oem_folder:
                 folder = api.content.create(container=oem_folder, type='Folder', id=uid, title=full_title)
-                alsoProvides(folder, IActionsPanelFolder)
+                alsoProvides(folder, IActionsPanelFolderOnlyAdd)
                 alsoProvides(folder, INextPrevNotNavigable)
                 roles = ['Reader']
                 if registry['imio.dms.mail.browser.settings.IImioDmsMailConfig.org_email_templates_encoder_can_edit']:
@@ -375,7 +376,7 @@ def contact_plonegroup_change(event):
             if uid not in cl_folder:
                 folder = api.content.create(container=cl_folder, type='Folder', id=uid, title=full_title)
                 folder.setLayout('folder_tabular_view')
-                alsoProvides(folder, IActionsPanelFolderAll)
+                alsoProvides(folder, IActionsPanelFolderOnlyAdd)
                 alsoProvides(folder, INextPrevNotNavigable)
                 roles = ['Reader', 'Contributor', 'Editor']
                 api.group.grant_roles(groupname='%s_encodeur' % uid, roles=roles, obj=folder)